requirements.txt
Flask==3.0.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

